Added a baseline Graph class.
The graph class represents a directed graph.
In the directed graph it represents a node having an edge point to another node.
A -> B, B -> C, A -> C, A -> A

In the Graph nodes are stored in a sorted order lexographically.
Nodes store the value of their name and their edge connections.
Nodes can have multiple unique edges to the same node.

Edges are stored in lexographical order based on the node they point to.