Added code to find the main method.
The CodeGraph file begins by using a source directory to find the Java file
that contains the main method.

By using BFS the algorithm searches the file structure checking all the Java
files in the directory to see if it contains the String "public static void main".
If it does that file is the one that contains the main method.

The pom file was also updated to change the Java compiler to be a minimum of Java 1.8.