Started work on String Encryption.
Strings are encrypted by converting each character to its byte code.
After that the bytes are converted to Integers in which they are
multiplied by 42 and then subtracted by 27000. Strings are replaced then
by the encrypted Integer array. A decryption method is also inserted
into the class file in order to decrypt the Strings at runtime.