Fixed bugs.
Fixed bugs for identifying strings and encrypting them. Fixed bugs for
identifying methods and classes. Fixed bugs for renaming methods and
classes.

Removed the ability to rename classes that implement interfaces since
the code would rename the implemented methods which causes errors.

Disabled renaming variables and parameters for now until the bugs are
fixed in there.