Fixed implements identification.
Originally all classes, no matter what, were being identified as classes
that did not implement an interface. That problem has been resolved.

There was also an issue that which flipped what should happen when a
class is identified as one that implements an interface. It has been
resolved so that classes that do not implement an interface now have
their methods identified and renamed while classes that do implement an
interface do not.