Started work on variable identification.
Created a new Variable Structure class in order to manage variables that
have been found.

Repurposed old variable renaming code in order to identify variables.
This is a rough start and nothing specific works too much just yet. Some
basic variable identification that kind of works right now. Miss
identifying things like 'return this.color' with 'return' as the
variable type and 'this.color' as the variable name. Will try to resolve
this for next commit.