Fixed identifaction issue and inner code remover.
When identifying interfaces it would also identify nested interfaces for
the ownership of the Codestructure when really it is a class or
interface that owns them. Now only the outermost interface is owned by
the codestructure.

Fixed the inner code remover so that it properly removes the inner code
of the curly braces.