Major overhaul of the ClassStructure
Refactoring the ClassStructure code to be similar to the newer
structures that have been implemented for the codegraph. This will allow
greater functionality for the class structure. Plan to remove the
outdated methods, soon the original renamer will be completely removed
in favor of the updated renamer that uses the code graph. This commit
was mostly focused on class identification and cleaning up the class
structure and code structure code.